refactor: 로또 생성 로직 LottoGenerator에 위임
