+++ src/main/java/model/Lottos.java
@@ -4,26 +4,13 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
 
 public class Lottos {
 
     private final List<Lotto> lottos;
 
-    private Lottos(List<Lotto> lottos) {
+    public Lottos(List<Lotto> lottos) {
         this.lottos = new ArrayList<>(lottos);
-    }
-
-    public static Lottos of(List<List<Integer>> manualLottoNumbers, int autoLottoAmount, LottoGenerator lottoGenerator) {
-        List<Lotto> lottos = Stream.concat(
-                manualLottoNumbers.stream()
-                        .map(Lotto::fromNumberList),
-                IntStream.range(0, autoLottoAmount)
-                        .mapToObj(i -> lottoGenerator.generateLotto())
-        ).collect(Collectors.toList());
-
-        return new Lottos(lottos);
     }
 
     public LottoResult calculateResult(WinningLotto winningLotto) {
